feat(cli): move helpers, improve index generation script
