+++ src/fancy/examples/text/text-rotate-stagger-demo.tsx
@@ -2,7 +2,7 @@ import { LayoutGroup, motion } from "motion/react"
 
 import TextRotate from "@/fancy/components/text/text-rotate"
 
-import { exampleImages } from "../_helpers/exampleImages"
+import { exampleImages } from "../../../utils/_helpers/exampleImages"
 
 export default function Preview() {
   return (
