+++ src/utils/_helpers/exampleImages.ts
